Edit IEmployeeFactory.cs to use long instead of int for the id

+++ EmployeeWebApp/Models/IEmployeeFactory.cs
@@ -2,10 +2,10 @@ namespace EmployeeWebApp.Models;
 
 public interface IEmployeeFactory
 {
-    Employee GetEmployeeById(int id);
+    Employee GetEmployeeById(long id);
     void AddEmployee(Employee employee);
     void UpdateEmployee(Employee employee);
-    void DeleteEmployee(int id);
+    void DeleteEmployee(long id);
     IEnumerable<Employee> GetEmployeeList();
     Task SeedInitialData();
 }
